Trait sorted_iter::one_or_less_iterator::OneOrLess

pub trait OneOrLess { }

Iterators that return either 1 or 0 item.

An iterator that returns nothing or one thing is always sorted.

More interestingly, one_or_less_iterator.flatten() is sorted if the inner iterator is itself sorted. This isn’t true of non-one_or_less iterators.